+++ lithops/job/job.py
@@ -26,17 +26,17 @@ from types import SimpleNamespace
 
 from lithops import utils
 from lithops.job.partitioner import create_partitions
-from lithops.storage.utils import create_func_key, create_data_key,\
+from lithops.storage.utils import create_func_key, create_data_key, \
     create_job_key, func_key_suffix
 from lithops.job.serialize import SerializeIndependent, create_module_data
-from lithops.constants import MAX_AGG_DATA_SIZE, LOCALHOST,\
+from lithops.constants import MAX_AGG_DATA_SIZE, LOCALHOST, \
     SERVERLESS, STANDALONE, CUSTOM_RUNTIME_DIR, FAAS_BACKENDS
 
 
 logger = logging.getLogger(__name__)
 
 FUNCTION_CACHE = set()
-
+MAX_DATA_IN_PAYLOAD = 8 * 1024  # Per invocation. 8KB
 
 def create_map_job(
     config,
@@ -258,10 +258,7 @@ def _create_job(
 
     # Upload function and data
     upload_function = not config['lithops'].get('customized_runtime', False)
-    upload_data = not (
-        (len(str(data_str)) * job.chunksize < 8 * 1204 for data_str in data_strs)
-        and backend in FAAS_BACKENDS
-    )
+    upload_data = any([(len(data_str) * job.chunksize) > MAX_DATA_IN_PAYLOAD for data_str in data_strs])
 
     # Upload function and modules
     if upload_function:
@@ -292,8 +289,8 @@ def _create_job(
         host_job_meta['host_func_upload_time'] = 0
 
     # upload data
-    if upload_data:
-        # Upload iterdata to COS only if a single element is greater than 8KB
+    if upload_data or backend not in FAAS_BACKENDS:
+        # Upload iterdata to COS only if a single element is greater than MAX_DATA_IN_PAYLOAD
         logger.debug('ExecutorID {} | JobID {} - Uploading data to the storage backend'
                      .format(executor_id, job_id))
         # pass_iteradata through an object storage file
@@ -310,7 +307,7 @@ def _create_job(
         # pass iteradata as part of the invocation payload
         logger.debug('ExecutorID {} | JobID {} - Data per activation is < '
                      '{}. Passing data through invocation payload'
-                     .format(executor_id, job_id, utils.sizeof_fmt(8 * 1024)))
+                     .format(executor_id, job_id, utils.sizeof_fmt(MAX_DATA_IN_PAYLOAD)))
         job.data_key = None
         job.data_byte_ranges = None
         job.data_byte_strs = data_strs